packages/tf-tts-audiomodel/requirements.txt
absl-py==0.15.0
astunparse==1.6.3
audioread==3.0.1
cached-property==1.5.2
cachetools==5.3.3
certifi==2024.2.2
cffi==1.15.1
charset-normalizer==3.3.2
clang==5.0
click==8.1.7
cycler==0.11.0
Cython==3.0.10
dataclasses==0.6
decorator==5.1.1
Distance==0.1.3
docopt==0.6.2
filelock==3.12.2
flatbuffers==1.12
fonttools==4.38.0
g2p-en==2.1.0
g2pM==0.1.2.5
gast==0.4.0
google-auth==2.29.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.62.2
h5py==3.1.0
huggingface-hub==0.0.8
idna==3.7
importlib-metadata==6.7.0
inflect==6.0.5
jamo==0.4.1
joblib==1.3.2
keras==2.6.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.5
librosa==0.9.2
llvmlite==0.31.0
Markdown==3.4.4
MarkupSafe==2.1.5
matplotlib==3.5.3
nltk==3.8.1
num2words==0.5.13
numba==0.48.0
numpy==1.19.5
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==24.0
Pillow==9.5.0
platformdirs==2.6.1
pooch==1.8.1
protobuf==3.20.3
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==1.9.2
pyparsing==3.1.2
pypinyin==0.51.0
python-dateutil==2.9.0.post0
pyworld==0.3.4
PyYAML==6.0.1
regex==2024.4.16
requests==2.31.0
requests-oauthlib==2.0.0
resampy==0.3.1
rsa==4.9
scikit-learn==1.0.2
scipy==1.7.3
six==1.15.0
soundfile==0.12.1
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-addons==0.19.0
tensorflow-estimator==2.15.0
tensorflow-gpu==2.6.0
TensorFlowTTS==1.8
termcolor==1.1.0
TextGrid==1.6.1
threadpoolctl==3.1.0
tqdm==4.66.2
typeguard==2.13.3
typing-extensions==3.7.4.3
Unidecode==1.3.8
urllib3==2.0.7
Werkzeug==2.2.3
wrapt==1.12.1
zipp==3.15.0
